+++ js/until.js
@@ -1,0 +1,32 @@
+function getRandomPositiveInteger (a, b) {
+  const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
+  const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+}
+
+function getRandomPositiveFloat (a, b, digits = 1) {
+  const lower = Math.min(Math.abs(a), Math.abs(b));
+  const upper = Math.max(Math.abs(a), Math.abs(b));
+  const result = Math.random() * (upper - lower) + lower;
+  return +result.toFixed(digits);
+}
+
+function getRamdonArrayEL(array) {
+  return array[getRandomPositiveInteger(1, array.length - 1)];
+}
+
+function shuffleArray(array) {
+  let j, temp;
+
+  for(let i = array.length - 1; i > 0; i--){
+    j = Math.floor(Math.random()*(i + 1));
+    temp = array[j];
+    array[j] = array[i];
+    array[i] = temp;
+  }
+
+  return array;
+}
+
+export {getRandomPositiveInteger, getRandomPositiveFloat, getRamdonArrayEL, shuffleArray};
